Changed: - pop_vector (vector passed as &) - BDD_ID as size_t - added missing terminal cases to ite - check terminal cases before computed_table - Removed label propagation for logical op. - cofactorTrue(f) and coFactorFalse(f) simplified

+++ src/ManagerInterface.h
@@ -10,7 +10,7 @@
 
 namespace ClassProject {
 
-    typedef int BDD_ID;
+    typedef size_t BDD_ID;
 
     class ManagerInterface {
     public:
